removed filename prefixes that redundantly reference the component name
