--- src/Nav/Nav.actions.js
@@ -1,5 +1,0 @@
-import * as types from './Nav.types';
-
-export const testAction = () => ({
-  type: types.HELLO,
-});
