--- src/Nav/Nav.menu.js
@@ -1,59 +1,0 @@
-import React from 'react';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
-import { withStyles } from '@material-ui/core/styles';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-import Button from '@material-ui/core/Button';
-import IconButton from '@material-ui/core/IconButton';
-import MenuIcon from '@material-ui/icons/Menu';
-
-import * as NavActions from './Nav.actions';
-
-
-const styles = {
-  root: {
-    flexGrow: 1,
-  },
-  flex: {
-    flex: 1,
-  },
-  menuButton: {
-    marginLeft: -12,
-    marginRight: 20,
-  },
-};
-
-function ButtonAppBar(props) {
-  const { classes, navActions } = props;
-  return (
-    <div className={classes.root}>
-      <AppBar position="static">
-        <Toolbar>
-          <IconButton className={classes.menuButton} color="inherit" aria-label="Menu">
-            <MenuIcon />
-          </IconButton>
-          <Typography variant="title" color="inherit" className={classes.flex}>
-            Caravan
-          </Typography>
-          <Button color="inherit" onClick={() => {
-            navActions.testAction();
-          }}>
-            Login
-          </Button>
-        </Toolbar>
-      </AppBar>
-    </div>
-  );
-}
-
-const mapStateToProps = state => ({
-  Nav: state.Nav,
-});
-
-const mapDispatchToProps = dispatch => ({
-  navActions: bindActionCreators(NavActions, dispatch),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(withStyles(styles)(ButtonAppBar));
