+++ src/Nav/actions.js
@@ -1,0 +1,5 @@
+import * as types from "./types";
+
+export const testAction = () => ({
+  type: types.HELLO
+});
