+++ src/Nav/reducer.js
@@ -1,15 +1,14 @@
-import * as types from './Nav.types';
-import InitialState from './Nav.initial-state';
+import * as types from "./types";
+import InitialState from "./initial-state";
 
 export default function NavReducer(state = InitialState, action) {
   switch (action.type) {
     case types.HELLO:
       return {
         ...state,
-        hello: true,
+        hello: true
       };
     default:
       return state;
   }
 }
-
